# exemple-service-api/src/test/resources/model/swagger.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "Api documentation",
		"description": "Api documentation",
		"version": "nc"
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"tags": [
		{
			"name": "account"
		},
		{
			"name": "login"
		},
		{
			"name": "stock"
		},
		{
			"name": "subscription"
		}
	],
	"paths": {
		"/ws/v1/accounts": {
			"post": {
				"tags": [
					"account"
				],
				"operationId": "create",
				"parameters": [
					{
						"in": "header",
						"schema": {
							"$ref": "#/components/schemas/SchemaBeanParam"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"oneOf": [
									{
										"$ref": "#/components/schemas/Account.v1.user"
									},
									{
										"$ref": "#/components/schemas/Account.v2.admin"
									}
								]
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Account is created",
						"headers": {
							"Location": {
								"description": "Links to Account Data",
								"style": "simple",
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer_authentification": []
					},
					{
						"oauth2_client_credentials": []
					}
				]
			}
		},
		"/ws/v1/accounts/{id}": {
			"get": {
				"tags": [
					"account"
				],
				"operationId": "get",
				"parameters": [
					{
						"in": "header",
						"schema": {
							"$ref": "#/components/schemas/SchemaBeanParam"
						}
					},
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Account Data",
						"content": {
							"application/json": {
								"schema": {
									"oneOf": [
										{
											"$ref": "#/components/schemas/Account.v1.user"
										},
										{
											"$ref": "#/components/schemas/Account.v2.admin"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "Account is not found"
					},
					"403": {
						"description": "Account is not accessible"
					}
				},
				"security": [
					{
						"bearer_authentification": []
					},
					{
						"oauth2_password": []
					}
				]
			},
			"put": {
				"tags": [
					"account"
				],
				"operationId": "update",
				"parameters": [
					{
						"in": "header",
						"schema": {
							"$ref": "#/components/schemas/SchemaBeanParam"
						}
					},
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"oneOf": [
									{
										"$ref": "#/components/schemas/Account.v1.user"
									},
									{
										"$ref": "#/components/schemas/Account.v2.admin"
									}
								]
							}
						}
					},
					"required": true
				},
				"responses": {
					"204": {
						"description": "Account is updated"
					},
					"404": {
						"description": "Account is not found"
					},
					"403": {
						"description": "Account is not accessible"
					}
				},
				"security": [
					{
						"bearer_authentification": []
					},
					{
						"oauth2_password": []
					}
				]
			},
			"patch": {
				"tags": [
					"account"
				],
				"operationId": "update_1",
				"parameters": [
					{
						"in": "header",
						"schema": {
							"$ref": "#/components/schemas/SchemaBeanParam"
						}
					},
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Patch"
							}
						}
					},
					"required": true
				},
				"responses": {
					"204": {
						"description": "Account is updated"
					},
					"404": {
						"description": "Account is not found"
					},
					"403": {
						"description": "Account is not accessible"
					}
				},
				"security": [
					{
						"bearer_authentification": []
					},
					{
						"oauth2_password": []
					}
				]
			}
		},
		"/ws": {
			"get": {
				"operationId": "template",
				"responses": {
					"default": {
						"description": "default response",
						"content": {
							"text/xml": {
								"schema": {
									"$ref": "#/components/schemas/Info"
								}
							},
							"text/html": {
								"schema": {
									"$ref": "#/components/schemas/Info"
								}
							}
						}
					}
				}
			}
		},
		"/ws/v1/logins/{username}": {
			"get": {
				"tags": [
					"login"
				],
				"operationId": "get_1",
				"parameters": [
					{
						"in": "header",
						"schema": {
							"$ref": "#/components/schemas/ApplicationBeanParam"
						}
					},
					{
						"name": "username",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Login Data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Login"
								}
							}
						}
					},
					"404": {
						"description": "Login is not found"
					},
					"403": {
						"description": "Login is not accessible"
					}
				},
				"security": [
					{
						"bearer_authentification": []
					},
					{
						"oauth2_password": []
					}
				]
			},
			"head": {
				"tags": [
					"login"
				],
				"operationId": "check",
				"parameters": [
					{
						"in": "header",
						"schema": {
							"$ref": "#/components/schemas/ApplicationBeanParam"
						}
					},
					{
						"name": "username",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"default": {
						"description": "default response",
						"content": {
							"*/*": {
								
							}
						}
					}
				},
				"security": [
					{
						"bearer_authentification": []
					},
					{
						"oauth2_client_credentials": []
					}
				]
			}
		},
		"/ws/v1/stocks/{store}/{product}": {
			"get": {
				"tags": [
					"stock"
				],
				"operationId": "get_2",
				"parameters": [
					{
						"in": "header",
						"schema": {
							"$ref": "#/components/schemas/ApplicationBeanParam"
						}
					},
					{
						"name": "store",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "product",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Stock Data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Stock"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer_authentification": []
					},
					{
						"oauth2_client_credentials": []
					}
				]
			},
			"post": {
				"tags": [
					"stock"
				],
				"operationId": "post",
				"parameters": [
					{
						"in": "header",
						"schema": {
							"$ref": "#/components/schemas/ApplicationBeanParam"
						}
					},
					{
						"name": "store",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "product",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Stock"
							}
						}
					},
					"required": true
				},
				"responses": {
					"default": {
						"description": "default response",
						"content": {
							"*/*": {
								"schema": {
									"type": "integer",
									"format": "int64"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer_authentification": []
					},
					{
						"oauth2_client_credentials": []
					}
				]
			}
		},
		"/ws/v1/subscriptions/{email}": {
			"get": {
				"tags": [
					"subscription"
				],
				"operationId": "get_3",
				"parameters": [
					{
						"in": "header",
						"schema": {
							"$ref": "#/components/schemas/SchemaBeanParam"
						}
					},
					{
						"name": "email",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Subscription Data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Subscription"
								}
							}
						}
					},
					"403": {
						"description": "Subscription is not accessible"
					}
				},
				"security": [
					{
						"bearer_authentification": []
					},
					{
						"oauth2_password": []
					}
				]
			},
			"put": {
				"tags": [
					"subscription"
				],
				"operationId": "update_2",
				"parameters": [
					{
						"in": "header",
						"schema": {
							"$ref": "#/components/schemas/SchemaBeanParam"
						}
					},
					{
						"name": "email",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Subscription"
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Subscription is created",
						"headers": {
							"Location": {
								"description": "Links to Subscription Data",
								"style": "simple",
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"204": {
						"description": "Subscription is updated",
						"headers": {
							"Location": {
								"description": "Links to Subscription Data",
								"style": "simple",
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer_authentification": []
					},
					{
						"oauth2_password": []
					}
				]
			}
		}
	},
	"components": {
		"schemas": {
			"SchemaBeanParam": {
				"required": [
					"app",
					"version"
				],
				"type": "object",
				"properties": {
					"app": {
						"type": "string"
					},
					"version": {
						"type": "string"
					}
				}
			},
			"Patch": {
				"$ref": "/ws/v1/schemas/patch"
			},
			"Info": {
				"type": "object",
				"properties": {
					"version": {
						"type": "string"
					},
					"buildTime": {
						"type": "string"
					}
				}
			},
			"ApplicationBeanParam": {
				"required": [
					"app"
				],
				"type": "object",
				"properties": {
					"app": {
						"type": "string"
					}
				}
			},
			"Stock": {
				"required": [
					"increment"
				],
				"type": "object",
				"properties": {
					"increment": {
						"type": "integer",
						"format": "int32",
						"writeOnly": true
					},
					"amount": {
						"type": "integer",
						"format": "int64",
						"readOnly": true
					}
				}
			},
			"Account.v1.user": {
				"$ref": "/ws/v1/schemas/account/test/v1/user"
			},
			"Account.v2.admin": {
				"$ref": "/ws/v1/schemas/account/test/v2/admin"
			}
		}
	}
}